+++ dist_app/app.js
@@ -32,6 +32,9 @@ var r = "skynovel", i = class {
 			setDec: (e) => {
 				this.dec = e;
 			},
+			setDecAB: (e) => {
+				this.decAB = e;
+			},
 			setEnc: (e) => {
 				this.enc = e;
 			},
@@ -75,6 +78,7 @@ var r = "skynovel", i = class {
 		return this.$path_userdata;
 	}
 	dec = (e, t) => Promise.resolve(t);
+	decAB = (e) => Promise.resolve(e);
 	enc = (e) => Promise.resolve(e);
 	fetch = (e, t) => fetch(e, t);
 	hash = (e) => "";